requirements.txt
# Voice-Based Intelligent Customer Service System
# Requirements File

# Core ML & NLP
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8.1
joblib>=1.3.0

# Deep Learning (Optional - for BERT)
transformers>=4.30.0
torch>=2.0.0

# Speech Processing
openai-whisper>=20230918
gTTS>=2.4.0
pyttsx3>=2.90
sounddevice>=0.4.6
scipy>=1.11.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Audio Processing
pydub>=0.25.1
soundfile>=0.12.1

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Generative AI
google-generativeai>=0.3.0

# Testing
pytest>=7.4.0
